+++ src/components/Project.vue
@@ -4,10 +4,10 @@
 			<h2>COMPARTILHE UM SONHO</h2>
 			<h3>{{ project.name }}</h3>
 			<header>
-				<div
-					class="project__thumbnail"
-					:style="{ backgroundImage: `url(${project.image_url})`}"
-				></div>
+				<div v-if="project.image_url || project.video_url" class="project__thumbnail" :class="'project__thumbnail--' + (project.video_url != null ? 'video' : 'image')">
+					<iframe v-if="project.video_url" :src="project.video_url" width="560" height="315" frameborder="0" allow="autoplay; encrypted-media" allowfullscreen></iframe>
+					<img v-else :src="project.image_url" :alt="project.image_alt">
+				</div>
 				<div class="project__header">
 					<p><strong>{{ project.summary.captured_count }}</strong> pessoas já ajudaram</p>
 					<p><strong>{{ getRemainingDays(project.end_ts) }}</strong> dias restantes</p>